Adicionando middleware default e do passport-github
