+++ src/index.js
@@ -6,17 +6,17 @@ const initializeDb = require('./database');
 const api = require('./api');
 const config = require('config');
 
-let app = express();
+const app = express();
 app.server = http.createServer(app);
 
 // logger
-app.use(morgan());
+app.use(morgan('dev'));
 
 // connect to db
 initializeDb( db => {
 
 	// internal middleware
-	app.use(middleware({ config, db }));
+	app.use(middleware({ config, db, app }));
 
 	// api router
 	app.use('/api', api({ config, db }));
