+++ src/middleware/index.js
@@ -1,93 +1,15 @@
 const { Router } = require('express');
-const { Strategy } = require('passport-github2');
-const passport = require('passport');
-const defaultMiddleware = require('./default-middleware');
+const defaultMiddleware = require('./default-middleware-application');
+const passportMiddleware = require('./passport-middleware');
 
 
-module.exports = ( { config, db } ) => {
+module.exports = ({ config, db, app }) => {
   console.info('Init Middleware module');
   const routes = Router();
-  //routes.use(defaultMiddleware);
 
-  const GITHUB_CLIENT_ID = "93c87ec9bfbea4151180";
-  const GITHUB_CLIENT_SECRET = "8e1d50fb000071ebd47a7c970fce9365c831f287";
+  defaultMiddleware({ app });
 
-
-  // Passport session setup.
-  //   To support persistent login sessions, Passport needs to be able to
-  //   serialize users into and deserialize users out of the session.  Typically,
-  //   this will be as simple as storing the user ID when serializing, and finding
-  //   the user by ID when deserializing.  However, since this example does not
-  //   have a database of user records, the complete GitHub profile is serialized
-  //   and deserialized.
-  passport.serializeUser(function(user, done) {
-    done(null, user);
-  });
-
-  passport.deserializeUser(function(obj, done) {
-    done(null, obj);
-  });
-
-
-  // Use the GitHubStrategy within Passport.
-  //   Strategies in Passport require a `verify` function, which accept
-  //   credentials (in this case, an accessToken, refreshToken, and GitHub
-  //   profile), and invoke a callback with a user object.
-  passport.use(new Strategy({
-      clientID: GITHUB_CLIENT_ID,
-      clientSecret: GITHUB_CLIENT_SECRET,
-      callbackURL: "http://127.0.0.1:3000/auth/github/callback"
-    },
-    function(accessToken, refreshToken, profile, done) {
-      // asynchronous verification, for effect...
-      process.nextTick(function () {
-
-        // To keep the example simple, the user's GitHub profile is returned to
-        // represent the logged-in user.  In a typical application, you would want
-        // to associate the GitHub account with a user record in your database,
-        // and return that user instead.
-        return done(null, profile);
-      });
-    }
-  ));
-
-
-
-  // Initialize Passport!  Also use passport.session() middleware, to support
-  // persistent login sessions (recommended).
-  routes.use(passport.initialize());
-  routes.use(passport.session());
-
-
-
-  // GET /auth/github
-  //   Use passport.authenticate() as route middleware to authenticate the
-  //   request.  The first step in GitHub authentication will involve redirecting
-  //   the user to github.com.  After authorization, GitHub will redirect the user
-  //   back to this application at /auth/github/callback
-  routes.get('/auth/github',
-    passport.authenticate('github', { scope: [ 'user:email' ] }),
-    function(req, res){
-      console.error('Erro');
-      // The request will be redirected to GitHub for authentication, so this
-      // function will not be called.
-    });
-
-
-  // GET /auth/github/callback
-  //   Use passport.authenticate() as route middleware to authenticate the
-  //   request.  If authentication fails, the user will be redirected back to the
-  //   login page.  Otherwise, the primary route function will be called,
-  //   which, in this example, will redirect the user to the home page.
-  routes.get('/auth/github/callback',
-    passport.authenticate('github', { failureRedirect: '/login' }),
-    function(req, res) {
-      console.log(req.user);
-      res.redirect('/api');
-    });
-
-
-	// add middleware here
+	routes.use(passportMiddleware({ config, db }));
 
 	return routes;
 }
